refactored workflow
