+++ skbio/workflow/tests/test_workflow.py
@@ -7,8 +7,8 @@
 # ----------------------------------------------------------------------------
 
 from collections import defaultdict
-from skbio.workflow import (Exists, NotExecuted, NotNone, Workflow, not_none,
-                            requires, method)
+from skbio.workflow._workflow import (
+    Exists, NotExecuted, NotNone, Workflow, not_none, requires, method)
 from unittest import TestCase, main
